Improved parser performance
The performance was improved by compressing the trie in a uint16 slice
that is generated with the parser instead of being created at runtime.
It was also modified the mechanism to find the first terminal token on
the stack, making it more efficient although a little less clear.

+++ languages/arithmetic/function.go
@@ -24,7 +24,7 @@ func parserPreallocMem(inputSize int, numThreads int) {
 /*
 function is the semantic function of the parser.
 */
-func function(thread int, ruleNum int, lhs *symbol, rhs []*symbol) {
+func function(thread int, ruleNum uint16, lhs *symbol, rhs []*symbol) {
 	switch ruleNum {
 	case 0:
 		NEW_AXIOM0 := lhs
